requirements.txt
Flask==2.2.2
Flask-Migrate==4.0.0
SQLAlchemy==1.4.44
flask-sqlalchemy==3.0.2
psycopg-binary==3.1.4
psycopg2==2.9.5
python-dotenv==0.21.0
pydantic==1.9.0
backoff==2.2.1
redis==4.3.5
flake8==5.0.4
gevent==22.10.2
gunicorn==20.1.0
argon2-cffi==21.3.0
PyJWT==2.6.0
flask-jwt-extended==4.4.4
Flask-RESTful==0.3.9
PyYAML==6.0

pytest==7.2.0
